# Conflicts:
#	src/main/java/com/ark/demo/controllers/RequestController.java
#	src/main/java/com/ark/demo/controllers/ResponseController.java
#	src/main/java/com/ark/demo/models/Request.java
#	src/main/java/com/ark/demo/models/User.java
#	src/main/resources/templates/requestTemplates/createRequest.html
#	src/main/resources/templates/requestTemplates/requestConfirmation.html
#	src/main/resources/templates/requestTemplates/userRequests.html

+++ src/main/java/com/ark/demo/controllers/UserController.java
@@ -21,6 +21,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.Optional;
 
 import static java.util.Objects.isNull;
 
@@ -40,11 +41,12 @@ public class UserController {
     public String displayUserProfile(HttpServletRequest request, Model model){
         User user = authenticationController.getUserFromSession(request.getSession());
         if(isNull(user)){
-            return "redirect:../login";
+            return "redirect:/login";
         }
         ViewProfileDTO viewProfileDTO = new ViewProfileDTO();
         viewProfileDTO.setUserDetails(user.getUserDetails());
         viewProfileDTO.setDateCreated(formatDateAsString(user.getDateCreated()));
+        model.addAttribute(user);
         model.addAttribute("title","View Profile");
         model.addAttribute(viewProfileDTO);
         return "userTemplates/viewProfile";
@@ -59,6 +61,7 @@ public class UserController {
         EditProfileFormDTO editProfileFormDTO = new EditProfileFormDTO();
         editProfileFormDTO.setUserDetails(user.getUserDetails());
         model.addAttribute(editProfileFormDTO);
+        model.addAttribute("states",authenticationController.createStatesMap());
         return "userTemplates/editProfile";
     }
 
@@ -84,6 +87,8 @@ public class UserController {
         userDetails.setZipcode(editProfileFormDTO.getUserDetails().getZipcode());
         userDetails.setPhoneNumber(editProfileFormDTO.getUserDetails().getPhoneNumber());
         userDetails.setEmailAddress(editProfileFormDTO.getUserDetails().getEmailAddress());
+        user.setLocation(editProfileFormDTO.getLocation());
+        userRepository.save(user);
         userDetailsRepository.save(userDetails);
         return "redirect:../user";
     }
@@ -140,6 +145,10 @@ public class UserController {
 
     @PostMapping("/deleteProfile")
     public String processDeleteProfileForm(@ModelAttribute @Valid DeleteFormDTO deleteFormDTO, HttpServletRequest request){
+        User user = authenticationController.getUserFromSession(request.getSession());
+        if(isNull(user)){
+            return "redirect:../login";
+        }
         if(deleteFormDTO.getConfirm().toLowerCase().equals("yes")){
             userDetailsRepository.delete(deleteFormDTO.getUserDetails());
             userRepository.delete(deleteFormDTO.getUser());
